requirements.txt
opencv-python==4.7.0.72
aiobotocore==2.13.0
onnxruntime==1.17.0
pillow==10.3.0
numpy==1.25.2
pyyaml==6.0.1